requirements.txt
# PyTorch for model inference (CPU/GPU)
torch>=1.12.0

# Ultralytics YOLOv5/YOLOv8 API
ultralytics>=8.0.0

# OpenCV для захвата и отображения видео
opencv-python>=4.5.0

# Чтение YAML-конфигураций
pyyaml>=6.0
